+++ tests/testthat/test-wflow_build.R
@@ -210,6 +210,35 @@ test_that("wflow_build accepts custom directory to save log files", {
   expect_identical(expected, actual$log_dir)
 })
 
+test_that("wflow_build automatically removes unused figure files", {
+  # Build a file that has 2 plots from 2 unnamed chunks
+  file_w_figs <- file.path(s$analysis, "fig.Rmd")
+  file.copy("files/test-wflow_build/figure-v01.Rmd", file_w_figs)
+  build_v01 <- wflow_build(file_w_figs, view = FALSE, project = site_dir)
+  figs_analysis_v01 <- file.path(s$analysis, "figure", basename(file_w_figs),
+                                 c("unnamed-chunk-1-1.png", "unnamed-chunk-2-1.png"))
+  expect_true(all(file.exists(figs_analysis_v01)))
+  figs_docs_v01 <- file.path(s$docs, "figure", basename(file_w_figs),
+                             c("unnamed-chunk-1-1.png", "unnamed-chunk-2-1.png"))
+  expect_true(all(file.exists(figs_docs_v01)))
+  # Update the file such that the previous 2 chunks are now named, plus add a
+  # 3rd plot chunk
+  file.copy("files/test-wflow_build/figure-v02.Rmd", file_w_figs, overwrite = TRUE)
+  build_v02 <- wflow_build(file_w_figs, view = FALSE, project = site_dir)
+  expect_false(all(file.exists(figs_analysis_v01)))
+  expect_false(all(file.exists(figs_docs_v01)))
+  figs_analysis_v02 <- file.path(s$analysis, "figure", basename(file_w_figs),
+                                 c("named1-1.png", "named2-1.png", "named3-1.png"))
+  expect_true(all(file.exists(figs_analysis_v02)))
+  figs_docs_v02 <- file.path(s$docs, "figure", basename(file_w_figs),
+                             c("named1-1.png", "named2-1.png", "named3-1.png"))
+  expect_true(all(file.exists(figs_docs_v02)))
+  # Cleanup
+  file.remove(file_w_figs)
+  unlink(file.path(s$analysis, "figure", basename(file_w_figs)), recursive = TRUE)
+  unlink(file.path(s$docs, "figure", basename(file_w_figs)), recursive = TRUE)
+})
+
 # Test error handling ----------------------------------------------------------
 
 test_that("wflow_build fails if file outside of analysis/", {